Starting a new project for our mathematician client: a web crawler
       \_______/        He has been thinking about graph theory a lot lately
   `.,-'\_____/`-.,'    and what the shape of the world wide web might be.
    /`..'\ _ /`.,'\     He wants a program that can visit every webpage that
   /  /`.,' `.,'\  \    has a link leading to it.
__/__/__/     \__\__\__
  \  \  \     /  /  /   I got excited when I found a few Python libraries
   \  \,'`._,'`./  /    that promised to make this very simple.  But then it
    \,'`./___\,'`./     started crashing.  A lot.  It turns out that there
   ,'`-./_____\,-'`.    are a lot of junk websites out there that don't
       /       \        follow the rules of the web.  My whole approach needs
                        to be re-imagined to deal with this.

There is one other wrinkle: our client insists that this       / _ \
program be RECURSIVE.  It has something to do with elegance  \_\(_)/_/
and simplicity.  I tried to make this into a recursive        _//"\\_  Max
program, but I've hated recursion ever since school and I      /   \
can't figure it out.  But the intern has been really good so
far... maybe they can figure it out for me.